DAOS-9883 libfabric: Apply two patches for libfabric (#53)
Revert patch with performance degradation and
apply the following patch for TCP provider.

Commit 45dce66 incorrectly negated a check that is used to signal
any threads waiting on a cq to wake up. Without this signal,
applications can hang, as reported in issue 7443. The situation is
as follows:

A thread waits for completions on the CQ. At the time that the thread
blocks, the socket has no outgoing transfers, so the thread calls
poll looking for a POLLIN event.

Another thread tries to send a large transfer. A portion of the data
is accepted by the socket. The rest of the transfer is queued on the ep.
That thread returns. Additional messages may be posted to the ep, but
since there's already an active transfer, those messages simply get
queued on the tx_queue.

Note that the thread that was blocked on the CQ is still blocked waiting
for a POLLIN event. However, it should be waiting for POLLOUT | POLLIN.
The signal in this case isn't trying to drive progress, like the original
commit suggests, but is needed to update the poll call.

Reverse the check, so that blocked threads wake up and can update their
poll event status.

Signed-off-by: Dmitry Eremin <dmitry.eremin@intel.com>