SRE-2832 Packaging update part 1
Packaging update part 1 which needs a test build of Libfabric.
No functional code changes.

In the future, there were be only Dockerfile.ubuntu which should
handle all the releases of Ubuntu that will be supported.

Currently the pbuilder program is refusing to validate HTTPS
certificates that are not using a well known root CA.

Signed-off-by: John E. Malmberg <john.malmberg@hpe.com>